Adjust file extension logic
This addresses a problem in the File model implementation where foundational rules are broken. The translation plugin forces itself upon its extended objects by using a Purgeable trait and dynamic property injection. This introduces complexity and makes logic difficult to trace for debugging. Plugins should optimally pull in the desired behavior, as opposed to it being pushed on to the stack with magic and contortion. The core should introduce functionality by extension, as is the norm

The old logic remains for v1.0 but has been reworked for v2.0. The Translatable Model now is contextually aware of attachments and replaces references to MLFile